requirements.txt
pg8000>=1.30.2
psycopg2>=2.9.9
PySide6>=6.6.2
PySide6_Addons>=6.6.2
PySide6_Essentials>=6.6.2
python-dateutil>=2.8.2
scramp>=1.4.4
shiboken6>=6.6.2
psutil
SQLAlchemy>=2.0.23
sqlalchemy-json>=0.7.0
python-dotenv==0.19.0
# Word templates + Jinja2
python-docx>=1.1.0
jinja2>=3.1.0
docxtpl>=0.16.0
openpyxl>=3.1.0
# Форматирование (опционально)
isort
black